Reorganize folder structure: separated admin and website components into distinct folders
