+++ app/Http/Controllers/Auth/AuthController.php
@@ -1,6 +1,7 @@
 <?php
+namespace App\Http\Controllers\Auth;
 
-namespace App\Http\Controllers;
+use App\Http\Controllers\Controller;
 use App\Models\User;
 use App\Http\Requests\UserRequest;
 use Auth;
@@ -10,8 +11,6 @@ use Validator;
 
 class AuthController extends Controller
 {
-
-
     public function login(Request $request)
     {
