+++ resources/views/components/alerts.blade.php
