+++ resources/views/components/auth_partials/default.blade.php
@@ -1,22 +1,21 @@
 <!DOCTYPE html>
 <html lang="en">
 
-@include('website.layouts.partials.head')
+@include('components.guest_partials.head')
 
 <body class="bg-gray-100 font-poppins text-gray-800">
 
     <div class="page">
 
         {{-- header --}}
-        @include('website.layouts.partials.header')
+        @include('components.auth_partials.header')
 
         <div>
-
             {{-- start main content --}}
             @yield('content')
         </div>
 
-        @include('website.layouts.partials.footer')
+        @include('components.guest_partials.footer')
 
     </div>
