+++ resources/views/components/guest_partials/footer.blade.php
