+++ resources/views/components/guest_partials/head.blade.php
