--- resources/views/website/users/show.blade.php
@@ -1,1 +1,0 @@
-e
